Adds CSS parsing
