+++ app/blog/render/middleware.js
@@ -18,8 +18,8 @@ var CACHE_CONTROL = "Cache-Control";
 
 const {minifyJS, minifyCSS} = require("./minify");
 const injectScreenshotScript = require("./injectScreenshotScript");
-const replaceFolderLinks = require("./replaceFolderLinks");
-const replaceFolderLinksCSS = require("./replaceFolderLinksCSS");
+const replaceFolderLinks = require("./replaceFolderLinks/html");
+const replaceFolderLinksCSS = require("./replaceFolderLinks/css");
 
 var cacheDuration = "public, max-age=31536000";
 var JS = "application/javascript";
